+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 -   **BREAKING**: drop AMD / Require.js / UMD support, use CommonJS style and assume a build process
 
+-   **BREAKING**: drop W3C-callback style, return `Promise`s
+
 
 ## v1.0.1 - 2015-02-24